after 1st beta

+++ application/models/freepark_model.php
@@ -11,8 +11,8 @@ class Freepark_model extends CI_Model {
     function __construct() {
         // Call the Model constructor
         parent::__construct();
-        //delete everything before today.
-        $q = "DELETE FROM freedays_tbl WHERE free_date < CURDATE()";
+        //delete everything before today  and any funny entries.
+        $q = "DELETE FROM freedays_tbl WHERE free_date < CURDATE() OR (owner='' AND userId='')";
         $query = $this->db->query($q);
     }
 
@@ -57,18 +57,29 @@ class Freepark_model extends CI_Model {
         $query = $this->db->query($q);
         foreach ($query->result_array() as $row) {
              $dd = new DateTime($row['free_date']);
-            $result[$dd->format('j')] = "***";
+            $result[$dd->format('j')] = "Requested";
         }
         return $result;
     }
     
     function reserve_available_date($user, $yymmdd) {
-         $q = "UPDATE freedays_tbl SET userId='". $user."' WHERE free_date LIKE '" . $yymmdd . "%' AND userId = '' LIMIT 1";
+        if (empty($user)) return;
+        //remove any request
+        $q = "DELETE FROM freedays_tbl WHERE free_date = '" . $yymmdd . "' AND owner = '' AND userId= '". $user."'";
         $this->db->query($q);
         if ($this->db->affected_rows() == 0) {
-            //add a reserve for this date
-            $q = "INSERT INTO freedays_tbl (userId, free_date) VALUES ('$user','$yymmdd');";
+            $q = "UPDATE freedays_tbl SET userId='". $user."' WHERE free_date = '" . $yymmdd . "' AND userId = '' LIMIT 1";
             $this->db->query($q);
+            if ($this->db->affected_rows() == 0) {
+                //if allocated to user, release.
+                 $q = "UPDATE freedays_tbl SET userId='' WHERE free_date = '" . $yymmdd . "' AND userId = '". $user."' LIMIT 1";
+                $this->db->query($q);
+                 if ($this->db->affected_rows() == 0) {
+                    //Ok reserve it
+                    $q = "INSERT INTO freedays_tbl (userId, free_date) VALUES ('$user','$yymmdd');";
+                $this->db->query($q);
+                 }
+            }
         }
     }
     
@@ -78,6 +89,7 @@ class Freepark_model extends CI_Model {
     }
 
     function do_free_for_owner($owner, $yymmdd, $bay) {
+        if (empty($owner)) return;
         //see if somebody has requested that date
         $q = "UPDATE freedays_tbl SET owner='". $owner."', parkId='".$bay."' WHERE  owner=''  and userId !='' and free_date LIKE '" . $yymmdd . "%' LIMIT 1";
         $this->db->query($q);
